Add linting and selectors to lint the code and centeralized the computed states using eslint,stylelint,prettier and reselect N/A
