+++ db.json
@@ -2,13 +2,23 @@
   "todos": [
     {
       "text": "hello",
-      "is_complete": false,
-      "id": 1
+      "is_complete": true,
+      "id": 4
     },
     {
-      "text": "new",
+      "text": "happy",
       "is_complete": true,
-      "id": 2
+      "id": 8
+    },
+    {
+      "text": "new todo",
+      "is_complete": true,
+      "id": 9
+    },
+    {
+      "text": "new todo",
+      "is_complete": false,
+      "id": 10
     }
   ],
   "contacts": [
