+++ src/reducers/index.js
@@ -1,4 +1,4 @@
-import {combineReducers} from 'redux';
+import { combineReducers } from 'redux';
 import login from './login';
 import todos from './todos';
 import contacts from './contacts';
@@ -11,9 +11,9 @@ const appReducer = combineReducers({
 
 const rootReducer = (state, action) => {
   if (action.type === 'LOGOUT_USER') {
-    state = undefined
+    state = undefined;
   }
-  return appReducer(state, action)
+  return appReducer(state, action);
 };
 
 export default rootReducer;
